+++ src/pages/Services/Store/components/DetailForm.tsx
@@ -24,11 +24,8 @@ export default function DetailForm({ form }: { form: FormInstance }) {
       </CustomForm.GridRow>
       <CustomForm.TextArea label="설명" name="description" maxLength={200} />
       <CustomForm.Input label="카테고리 목록" name="category_ids" disabled />
-
       <StoreCategory form={form} />
-
       <OpenTimeForm form={form} />
-
       <Divider orientation="left" style={{ marginTop: '40px' }}>
         옵션
       </Divider>
@@ -39,7 +36,6 @@ export default function DetailForm({ form }: { form: FormInstance }) {
           </CustomForm.Checkbox>
         ))}
       </S.CheckboxWrap>
-
       <Divider orientation="left">사진</Divider>
       <S.UploadWrap>
         <CustomForm.MultipleUpload domain="lands" name="image_urls" form={form} />
